+++ src/main/resources/application.properties
@@ -1,3 +1,7 @@
 fun7.o7tools.api.base-url=https://us-central1-o7tools.cloudfunctions.net
 fun7.o7tools.api.username=fun7user
 fun7.o7tools.api.password=fun7pass
+
+# MONGODB (MongoProperties)
+spring.data.mongodb.uri=mongodb://localhost:27017/fun7
+spring.data.mongodb.auto-index-creation=false
