+++ src/test/resources/application-test.properties
@@ -1,6 +1,7 @@
 # MONGODB (MongoProperties)
-spring.data.mongodb.uri=mongodb://localhost:1/fun7
+spring.data.mongodb.database=test
+spring.data.mongodb.port=27017
 spring.mongodb.embedded.version=7.0
 
-fun7.o7tools.api.base-url=http://localhost:\${stub.server.port}
 stub.server.port=9561
+fun7.o7tools.api.base-url=http://localhost:\${stub.server.port}
